Fix warning in CI
As far as I could tell, it would have been compiled to a zero.
I think changing it to non zero to block other event lowlevel handlers
is the right decision and may fix some bugs, though I am not on Windows
and am unable to test any of this.